Moved and renamed fetchData method from Storage - Changed to parseInput in Parser
